+++ lib/units/device/index.js
@@ -21,7 +21,6 @@ module.exports = function(options) {
         .dependency(require('./plugins/screen/stream'))
         .dependency(require('./plugins/screen/capture'))
         .dependency(require('./plugins/service'))
-        .dependency(require('./plugins/display'))
         .dependency(require('./plugins/browser'))
         .dependency(require('./plugins/store'))
         .dependency(require('./plugins/clipboard'))
@@ -34,7 +33,6 @@ module.exports = function(options) {
         .dependency(require('./plugins/cleanup'))
         .dependency(require('./plugins/reboot'))
         .dependency(require('./plugins/connect'))
-        .dependency(require('./plugins/data'))
         .dependency(require('./plugins/account'))
         .dependency(require('./plugins/ringer'))
         .dependency(require('./plugins/wifi'))
