+++ lib/units/device/plugins/account.js
@@ -6,7 +6,7 @@ var wireutil = require('../../../wire/util')
 
 module.exports = syrup.serial()
   .dependency(require('./service'))
-  .dependency(require('./identity'))
+  .dependency(require('./util/identity'))
   .dependency(require('./touch'))
   .dependency(require('../support/router'))
   .dependency(require('../support/push'))
