+++ lib/units/device/plugins/group.js
@@ -11,7 +11,7 @@ var lifecycle = require('../../../util/lifecycle')
 
 module.exports = syrup.serial()
   .dependency(require('./solo'))
-  .dependency(require('./identity'))
+  .dependency(require('./util/identity'))
   .dependency(require('./service'))
   .dependency(require('../support/router'))
   .dependency(require('../support/push'))
