+++ lib/units/device/plugins/screen/capture.js
@@ -13,7 +13,7 @@ module.exports = syrup.serial()
   .dependency(require('../../support/push'))
   .dependency(require('../../support/storage'))
   .dependency(require('../../resources/minicap'))
-  .dependency(require('../display'))
+  .dependency(require('../util/display'))
   .define(function(options, adb, router, push, storage, minicap, display) {
     var log = logger.createLogger('device:plugins:screen:capture')
     var plugin = Object.create(null)
