+++ lib/units/device/plugins/screen/stream.js
@@ -21,7 +21,7 @@ var FailCounter = require('../../../../util/failcounter')
 module.exports = syrup.serial()
   .dependency(require('../../support/adb'))
   .dependency(require('../../resources/minicap'))
-  .dependency(require('../display'))
+  .dependency(require('../util/display'))
   .dependency(require('./options'))
   .define(function(options, adb, minicap, display, screenOptions) {
     var log = logger.createLogger('device:plugins:screen:stream')
