+++ lib/units/device/plugins/solo.js
@@ -10,7 +10,7 @@ module.exports = syrup.serial()
   .dependency(require('../support/sub'))
   .dependency(require('../support/push'))
   .dependency(require('../support/router'))
-  .dependency(require('./identity'))
+  .dependency(require('./util/identity'))
   .define(function(options, sub, push, router, identity) {
     var log = logger.createLogger('device:plugins:solo')
