+++ lib/units/device/plugins/touch/index.js
@@ -19,7 +19,7 @@ module.exports = syrup.serial()
   .dependency(require('../../support/adb'))
   .dependency(require('../../support/router'))
   .dependency(require('../../resources/minitouch'))
-  .dependency(require('../flags'))
+  .dependency(require('../util/flags'))
   .define(function(options, adb, router, minitouch, flags) {
     var log = logger.createLogger('device:plugins:touch')
