+++ lib/units/device/plugins/util/data.js
@@ -1,7 +1,7 @@
 var syrup = require('stf-syrup')
 var deviceData = require('stf-device-db')
 
-var logger = require('../../../util/logger')
+var logger = require('../../../../util/logger')
 
 module.exports = syrup.serial()
   .dependency(require('./identity'))
