+++ lib/units/device/plugins/util/display.js
@@ -3,14 +3,14 @@ var util = require('util')
 var syrup = require('stf-syrup')
 var EventEmitter = require('eventemitter3').EventEmitter
 
-var logger = require('../../../util/logger')
-var streamutil = require('../../../util/streamutil')
+var logger = require('../../../../util/logger')
+var streamutil = require('../../../../util/streamutil')
 
 module.exports = syrup.serial()
-  .dependency(require('../support/adb'))
-  .dependency(require('../resources/minicap'))
-  .dependency(require('./service'))
-  .dependency(require('./screen/options'))
+  .dependency(require('../../support/adb'))
+  .dependency(require('../../resources/minicap'))
+  .dependency(require('../service'))
+  .dependency(require('../screen/options'))
   .define(function(options, adb, minicap, service, screenOptions) {
     var log = logger.createLogger('device:plugins:display')
