+++ lib/units/device/plugins/util/identity.js
@@ -1,10 +1,10 @@
 var syrup = require('stf-syrup')
 
-var devutil = require('../../../util/devutil')
-var logger = require('../../../util/logger')
+var devutil = require('../../../../util/devutil')
+var logger = require('../../../../util/logger')
 
 module.exports = syrup.serial()
-  .dependency(require('../support/properties'))
+  .dependency(require('../../support/properties'))
   .dependency(require('./display'))
   .dependency(require('./phone'))
   .define(function(options, properties, display, phone) {
