+++ lib/units/device/plugins/util/phone.js
@@ -1,9 +1,9 @@
 var syrup = require('stf-syrup')
 
-var logger = require('../../../util/logger')
+var logger = require('../../../../util/logger')
 
 module.exports = syrup.serial()
-  .dependency(require('./service'))
+  .dependency(require('../service'))
   .define(function(options, service) {
     var log = logger.createLogger('device:plugins:phone')
