+++ lib/units/device/plugins/util/urlformat.js
@@ -3,8 +3,8 @@ var _ = require('lodash')
 var tr = require('transliteration')
 
 module.exports = syrup.serial()
-  .dependency(require('../identity'))
-  .dependency(require('../data'))
+  .dependency(require('./identity'))
+  .dependency(require('./data'))
   .define(function(options, identity, data) {
     function createSlug() {
       var model = identity.model
